refactor: move internal/local/local to internal/service
