+++ internal/cmd/local/credentials.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 
 	"github.com/airbytehq/abctl/internal/airbyte"
-	"github.com/airbytehq/abctl/internal/cmd/local/local"
 	"github.com/airbytehq/abctl/internal/k8s"
+	"github.com/airbytehq/abctl/internal/service"
 	"github.com/airbytehq/abctl/internal/telemetry"
 	"github.com/pterm/pterm"
 	"go.opencensus.io/trace"
@@ -33,7 +33,7 @@ func (cc *CredentialsCmd) Run(ctx context.Context, provider k8s.Provider, telCli
 	spinner := &pterm.DefaultSpinner
 
 	return telClient.Wrap(ctx, telemetry.Credentials, func() error {
-		k8sClient, err := local.DefaultK8s(provider.Kubeconfig, provider.Context)
+		k8sClient, err := service.DefaultK8s(provider.Kubeconfig, provider.Context)
 		if err != nil {
 			pterm.Error.Println("No existing cluster found")
 			return nil
